requirements.txt
blinker==1.8.2
gunicorn==22.0.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy~=1.26.4
packaging==24.1
paho-mqtt==1.6.1
pillow==10.3.0
pytesseract==0.3.10
Werkzeug==3.0.3
joblib~=1.4.2
pandas~=2.2.2
spacy~=3.7.5
scikit-learn~=1.5.0
spacy[en_core_web_sm]
opencv-python-headless
